back/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pymongo==4.8.0
google-generativeai==0.7.2
python-multipart==0.0.9
Pillow==10.4.0
